Rebase from 'debian/melodic/actionlib'

+++ include/actionlib/client/simple_action_client.h
@@ -64,7 +64,7 @@ namespace actionlib
 /**
  * \brief A Simple client implementation of the ActionInterface which supports only one goal at a time
  *
- * The SimpleActionClient wraps the exisitng ActionClient, and exposes a limited set of easy-to-use hooks
+ * The SimpleActionClient wraps the existing ActionClient, and exposes a limited set of easy-to-use hooks
  * for the user. Note that the concept of GoalHandles has been completely hidden from the user, and that
  * they must query the SimplyActionClient directly in order to monitor a goal.
  */
@@ -601,7 +601,8 @@ bool SimpleActionClient<ActionSpec>::waitForResult(const ros::Duration & timeout
       time_left = loop_period;
     }
 
-    done_condition_.timed_wait(lock, boost::posix_time::milliseconds(time_left.toSec() * 1000.0f));
+    done_condition_.timed_wait(lock,
+      boost::posix_time::milliseconds(static_cast<int64_t>(time_left.toSec() * 1000.0f)));
   }
 
   return cur_simple_state_ == SimpleGoalState::DONE;
